fix: take NativeCULong off the public surface (BL-067)
The 96 CKF/CK/CKZ constants were `static readonly NativeCULong`, a marshalling
type from a separate package whose width differs per RID. Three costs: consumers
took a hard public dependency on an interop package they otherwise never touch,
any breaking change there became a breaking change here, and nothing composed
with this library's own ulong-typed flags without `.Value` - which the library
itself wrote 68 times, once per flag predicate. They are now `public const
ulong`; const rather than static readonly, so they also work in case labels and
attribute arguments.

[Flags] enums were the other option in the entry and are the wrong shape for
this class of constant. CKF is a mixed bit space where one value means different
things by context - CKF_TOKEN_PRESENT, CKF_RNG and CKF_HW are all 0x1 - so it
cannot be one enum, and splitting it per context would duplicate the typed view
that SlotFlags/TokenFlags/SessionFlags/MechanismFlags/InterfaceFlags/
LibraryFlags already provide. An enum would also have re-added the very cast
this removes, since those records expose `ulong Flags`.

Converting only the constants would have left the stated goal unmet, because the
13 CK*Extensions classes still published 26 `ToCULong`/`ToCK*` methods over
NativeCULong. Nothing outside Native/, MechanismParams/ and Internal/ calls
them, so they are internal now, and the six flag records take `ulong` in their
internal constructors. NativeCULong is consequently absent from the public
surface: seven test files dropped `using KerckhoffsLabs.Runtime.InteropServices`
outright, which is the consumer experience in miniature.

CK_UNAVAILABLE_INFORMATION stays static readonly and still varies by platform.
That is the spec's definition, not an oversight - all bits set in a CK_ULONG,
32-bit on Windows and pointer-sized elsewhere - and token-reported counts widen
to ulong verbatim, so the sentinel has to match that width to be comparable at
all. It now derives from the marshalled width rather than an
OperatingSystem.IsWindows() branch, which is additionally correct on 32-bit
Unix, and the field says why.

PublicSurfaceDependencyTests walks every exported type's fields, parameters,
return types, bases and interfaces, unwrapping by-ref/array/generic decoration,
and fails on any type from the interop assembly. It was checked against a
deliberate regression (re-publicising CKRExtensions) rather than assumed, and a
companion test asserts the marshalling layer still holds NativeCULong fields, so
an empty offender list cannot quietly mean the package left the build.

Internal call sites came out simpler, not noisier: the cancel-flag accumulation
in Pkcs11Session is now `ulong cancelFlags |= CKF.CKF_DIGEST` instead of casting
to ulong and back on every line.

+++ src/KerckhoffsLabs.Security.Cryptography.Pkcs11.Tests/Unit/MechanismParams/BuildMarshalableTests.cs
@@ -173,7 +173,7 @@ public sealed class BuildMarshalableTests
         Assert.Equal((ulong)CKG.CKG_MGF1_SHA256, (ulong)s.Mgf);
         // Hardcoded, not caller-supplied: a change here would mean the struct stopped declaring that
         // the source data is present, and the token would ignore the buffer.
-        Assert.Equal((ulong)CKZ.CKZ_DATA_SPECIFIED, (ulong)s.Source);
+        Assert.Equal(CKZ.CKZ_DATA_SPECIFIED, (ulong)s.Source);
         Assert.Equal(2UL, (ulong)s.SourceDataLen);
         AssertBlockHolds(s.SourceData, [0x0A, 0x0B]);
     }
